Add files via upload
I used the NumPy library which makes it relatively easy to implement vectorized solutions of Monte Carlo simulations such as this. I then used matplotlib to create and display a plot showing each of the coordinates generated as a dot in the unit square using different colors for the points within the unit circle and the points outside the unit circle.