Add files via upload
Using a combination of Pandas, NumPy, and native Python analyze a data file `P4-ScoreData.csv` which contains results for a quiz for which you will need to calculate the  _Discrimination_ statistic for each question, calculating a correlation between test questions and overall performance. Correlation may be applied to exam and quiz question results to determine how effective each question is at differentiating between students that did well and students that did not do well on the assessment.